Adding the order total field. Some changes.

+++ OnlineStore.MVC/Models/Coupon/CouponViewModel.cs
@@ -12,7 +12,7 @@
 
         public DateTimeOffset? FinishDate { get; set; }
 
-        public double DiscountSize { get; set; }
+        public decimal DiscountSize { get; set; }
 
         public double PercentDiscountSize { get; set; }
