+++ OnlineStore.MVC/Models/Product/CreateProductViewModel.cs
@@ -4,9 +4,9 @@
     {
         public string? Name { get; set; }
 
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
 
         public int UnitsInStock { get; set; }
